Add token test

+++ serializers/token.go
@@ -1,7 +1,7 @@
 package serializers
 
 type TokenResponse struct {
-	Id           int    `jsonapi:"primary,token"`
+	Id           int64    `jsonapi:"primary,token"`
 	AccessToken  string `jsonapi:"attr,access_token"`
 	ExpiresIn    int64  `jsonapi:"attr,expires_in"`
 	RefreshToken string `jsonapi:"attr,refresh_token"`
